+++ example/apollo-server/movies.js
@@ -1,6 +1,6 @@
 import { augmentTypeDefs, augmentSchema } from '../../src/index';
 import { ApolloServer, gql, makeExecutableSchema } from 'apollo-server';
-import { v1 as neo4j } from 'neo4j-driver';
+import neo4j from 'neo4j-driver';
 import { typeDefs, resolvers } from './movies-schema';
 
 const schema = makeExecutableSchema({